Refactor marker bits into separate module.
