+++ gmaps/__init__.py
@@ -1,6 +1,7 @@
 from ._version import version_info, __version__
 
 from .maps import *
+from .marker import *
 
 def _jupyter_nbextension_paths():
     return [{
